refactor(cli): import, promote, export and import-knowledge move to src/cli/knowledge.ts
